LayoutTests/http/tests/loading/authentication-after-redirect-stores-wrong-credentials/authentication-after-redirect-stores-wrong-credentials.html

<html>
<script>
if (window.layoutTestController) {
	layoutTestController.dumpAsText();
	layoutTestController.dumpChildFramesAsText();
	layoutTestController.waitUntilDone();
	layoutTestController.setHandlesAuthenticationChallenges(true);
    layoutTestController.setAuthenticationUsername("httpUsername");
    layoutTestController.setAuthenticationPassword("httpPassword");
}
</script>

<body>
This test causes an HTTP authentication challenge in the middle of a redirect chain.<br>
It then loads a new page after that challenge, and the new page should *not* get any credentials passed to it.<br>
See <a href="https://bugs.webkit.org/show_bug.cgi?id=78003">bug 78003</a> for more details.<br>
<iframe src="http://localhost:8000/loading/authentication-after-redirect-stores-wrong-credentials/resources/wrong-credential-1-redirect-to-auth.php">
</iframe><br>
</body>
</html>
